采购进退货，销售售退货全部关联库存，样式重复叠加问题解决（$css.bind）

+++ src/main/java/com/sunjiamin/invoice/model/Purchase.java
@@ -24,7 +24,7 @@ public class Purchase implements Serializable {
 	@Column(name = "price")
 	private double price;
 	@Column(name = "count")
-	private double count;
+	private int count;
 	@Column(name = "totalPrice")
 	private double totalPrice;
 	@Column(name = "time")
@@ -62,11 +62,11 @@ public class Purchase implements Serializable {
 		this.price = price;
 	}
 
-	public double getCount() {
+	public int getCount() {
 		return count;
 	}
 
-	public void setCount(double count) {
+	public void setCount(int count) {
 		this.count = count;
 	}
